pyproject.toml
[project]
name = "ai-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.9",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
]
